tests/wpt/web-platform-tests/2dcontext/line-styles/2d.line.join.miter.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>Canvas test: 2d.line.join.miter</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/2dcontext/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/2dcontext/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.line.join.miter</h1>
<p class="desc">lineJoin 'miter' is rendered correctly</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="/images/green-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("lineJoin 'miter' is rendered correctly");
_addTest(function(canvas, ctx) {

ctx.fillStyle = '#0f0';
ctx.fillRect(0, 0, 100, 50);

ctx.lineJoin = 'miter';
ctx.lineWidth = 20;

ctx.fillStyle = '#f00';
ctx.strokeStyle = '#0f0';

ctx.fillStyle = '#f00';
ctx.strokeStyle = '#0f0';

ctx.fillRect(10, 10, 30, 20);
ctx.fillRect(20, 10, 20, 30);

ctx.beginPath();
ctx.moveTo(10, 20);
ctx.lineTo(30, 20);
ctx.lineTo(30, 40);
ctx.stroke();


ctx.fillStyle = '#0f0';
ctx.strokeStyle = '#f00';

ctx.beginPath();
ctx.moveTo(60, 20);
ctx.lineTo(80, 20);
ctx.lineTo(80, 40);
ctx.stroke();

ctx.fillRect(60, 10, 30, 20);
ctx.fillRect(70, 10, 20, 30);

_assertPixel(canvas, 38,12, 0,255,0,255, "38,12", "0,255,0,255");
_assertPixel(canvas, 39,11, 0,255,0,255, "39,11", "0,255,0,255");
_assertPixel(canvas, 40,10, 0,255,0,255, "40,10", "0,255,0,255");
_assertPixel(canvas, 41,9, 0,255,0,255, "41,9", "0,255,0,255");
_assertPixel(canvas, 42,8, 0,255,0,255, "42,8", "0,255,0,255");

_assertPixel(canvas, 88,12, 0,255,0,255, "88,12", "0,255,0,255");
_assertPixel(canvas, 89,11, 0,255,0,255, "89,11", "0,255,0,255");
_assertPixel(canvas, 90,10, 0,255,0,255, "90,10", "0,255,0,255");
_assertPixel(canvas, 91,9, 0,255,0,255, "91,9", "0,255,0,255");
_assertPixel(canvas, 92,8, 0,255,0,255, "92,8", "0,255,0,255");


});
</script>
